pyproject.toml
[project]
name = "calibration-python-und-akustik"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "jupyter",
    "spectacoular[full]",
]
